Massively simplified architecture

+++ db/migrate/20160703103433_create_scores.rb
@@ -2,7 +2,7 @@ class CreateScores < ActiveRecord::Migration
   def change
     create_table :scores do |t|
       t.references :agent, index: true, foreign_key: true
-      t.integer :value
+      t.float :value
       t.integer :week
 
       t.timestamps null: false
